requirements.txt
flake8~=7.1.1
flake8-bugbear~=24.10.31
flake8-black~=0.3.6
black~=24.4.2
requests~=2.32.3
pyyaml~=6.0.3